+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Changed
   - Use bumpver current_version instead of pep440_version. (GitHub #30)
 
+- Removed
+  - Removed Google Container Registry support.  (GitHub #29)
+
 ## Release History
 
 ### [38.0.0] - 2025-04-12